feat/Added enable view option in ERD
